requirements.txt
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
executing==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
ipython==9.2.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysqlclient==2.2.7
parso==0.8.4
pillow==11.2.1
prettytable==3.16.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
